requirements.txt
beautifulsoup4==4.13.3
colorama==0.4.6
dateparser==1.2.1
markdownify==1.1.0
matplotlib==3.10.0
numpy==2.2.3
openai==1.65.3
opencv-python==4.11.0.86
pdfplumber==0.11.5
pikepdf==9.5.2
pillow==11.1.0
PyMuPDF==1.25.3
pypandoc==1.15
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
tabulate==0.9.0
